test(coderd/x/nats): consolidate pubsub tests
